Lazy load inbox metadata scraper
