+++ apps/desktop/src/main/inbox/metadata.ts
@@ -21,6 +21,9 @@ import { createWriteStream } from 'fs'
 import { mkdir } from 'fs/promises'
 import { pipeline } from 'stream/promises'
 import { join, extname } from 'path'
+import { isBotPageTitle } from './metadata-utils'
+
+export { extractDomain, isBotPageTitle, titleFromUrl } from './metadata-utils'
 
 const log = createLogger('Inbox:Metadata')
 
@@ -306,54 +309,5 @@ export function isValidUrl(str: string): boolean {
     return url.protocol === 'http:' || url.protocol === 'https:'
   } catch {
     return false
-  }
-}
-
-import { extractDomain as extractDomainNullable } from '../lib/url-utils'
-
-export function extractDomain(url: string): string {
-  return extractDomainNullable(url) ?? url
-}
-
-const BOT_PAGE_TITLES = [
-  'just a moment...',
-  'attention required!',
-  'access denied',
-  'please wait',
-  'verify you are human',
-  'checking your browser'
-]
-
-export function isBotPageTitle(title: string): boolean {
-  if (!title) return false
-  const lower = title.toLowerCase()
-  return BOT_PAGE_TITLES.some((bot) => lower.includes(bot))
-}
-
-export function titleFromUrl(url: string): string {
-  try {
-    const parsed = new URL(url)
-    const segments = parsed.pathname.split('/').filter(Boolean)
-
-    if (segments.length === 0) {
-      return parsed.hostname.replace(/^www\./, '')
-    }
-
-    const last = segments[segments.length - 1]
-
-    const cleaned = last
-      .replace(/\.[a-z]+$/, '')
-      .replace(/--\d+$/, '')
-      .replace(/-\d+$/, '')
-      .replace(/[-_]+/g, ' ')
-      .trim()
-
-    if (!cleaned) {
-      return parsed.hostname.replace(/^www\./, '')
-    }
-
-    return cleaned.replace(/\b\w/g, (c) => c.toUpperCase())
-  } catch {
-    return url
   }
 }
